+++ src/aeif_psc_delta_rk5.h
@@ -20,20 +20,20 @@ struct aeif_psc_delta_rk5;
 
 template<int NVAR, int NPARAM>
 __device__
-void Derivatives(float x, float *y, float *dydx, float *param,
+void Derivatives(double x, float *y, float *dydx, float *param,
 		 aeif_psc_delta_rk5 data_struct);
 
 template<int NVAR, int NPARAM>
 __device__
-void ExternalUpdate(float x, float *y, float *param, bool end_time_step,
+void ExternalUpdate(double x, float *y, float *param, bool end_time_step,
 		    aeif_psc_delta_rk5 data_struct);
 
 __device__
-void NodeInit(int n_var, int n_param, float x, float *y,
+void NodeInit(int n_var, int n_param, double x, float *y,
 	      float *param, aeif_psc_delta_rk5 data_struct);
 
 __device__
-void NodeCalibrate(int n_var, int n_param, float x, float *y,
+void NodeCalibrate(int n_var, int n_param, double x, float *y,
 		   float *param, aeif_psc_delta_rk5 data_struct);
 
 #endif
