+++ src/iaf_psc_exp_g.cu
@@ -141,7 +141,7 @@ int iaf_psc_exp_g::Init(int i_node_0, int n_node, int /*n_port*/,
   return 0;
 }
 
-int iaf_psc_exp_g::Update(int it, float t1)
+int iaf_psc_exp_g::Update(long long it, double t1)
 {
   // std::cout << "iaf_psc_exp_g neuron update\n";
   float h = time_resolution_;
