+++ src/spike_detector.h
@@ -31,7 +31,7 @@ class spike_detector : public BaseNeuron
 
   int Free();
   
-  int Update(int it, float t1);
+  int Update(long long it, double t1);
 
 };
