+++ src/user_m1.cu
@@ -23,7 +23,7 @@ namespace user_m1_ns
 {
 
 __device__
-void NodeInit(int n_var, int n_param, float x, float *y, float *param,
+void NodeInit(int n_var, int n_param, double x, float *y, float *param,
 	      user_m1_rk5 data_struct)
 {
   //int array_idx = threadIdx.x + blockIdx.x * blockDim.x;
@@ -56,7 +56,7 @@ void NodeInit(int n_var, int n_param, float x, float *y, float *param,
 }
 
 __device__
-void NodeCalibrate(int n_var, int n_param, float x, float *y,
+void NodeCalibrate(int n_var, int n_param, double x, float *y,
 		       float *param, user_m1_rk5 data_struct)
 {
   //int array_idx = threadIdx.x + blockIdx.x * blockDim.x;
@@ -89,14 +89,14 @@ void NodeCalibrate(int n_var, int n_param, float x, float *y,
 }
 			    
 __device__
-void NodeInit(int n_var, int n_param, float x, float *y,
+void NodeInit(int n_var, int n_param, double x, float *y,
 	     float *param, user_m1_rk5 data_struct)
 {
     user_m1_ns::NodeInit(n_var, n_param, x, y, param, data_struct);
 }
 
 __device__
-void NodeCalibrate(int n_var, int n_param, float x, float *y,
+void NodeCalibrate(int n_var, int n_param, double x, float *y,
 		  float *param, user_m1_rk5 data_struct)
 
 {
@@ -144,7 +144,7 @@ int user_m1::Init(int i_node_0, int n_node, int n_port,
   return 0;
 }
 
-int user_m1::Calibrate(float time_min, float /*time_resolution*/)
+int user_m1::Calibrate(double time_min, float /*time_resolution*/)
 {
   rk5_.Calibrate(time_min, h_, rk5_data_struct_);
   
@@ -152,12 +152,12 @@ int user_m1::Calibrate(float time_min, float /*time_resolution*/)
 }
 
 template <>
-int user_m1::UpdateNR<0>(int it, float t1)
+int user_m1::UpdateNR<0>(long long it, double t1)
 {
   return 0;
 }
 
-int user_m1::Update(int it, float t1) {
+int user_m1::Update(long long it, double t1) {
   UpdateNR<MAX_PORT_NUM>(it, t1);
 
   return 0;
